backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.7.1

# Database & Utilities
sqlalchemy==2.0.30
python-dotenv==1.0.0
python-multipart==0.0.9

# PDF/DOCX Processing
pdfplumber==0.10.0
python-docx==1.1.0

# --- THE DEFINITIVE FIX: A Stable AI Stack ---
spacy==3.7.4
torch==2.2.1
transformers==4.41.2
# We are now using sentence-transformers directly with a powerful model
sentence-transformers==2.7.0
# huggingface-hub is a dependency of the above, this ensures a compatible version
huggingface-hub==0.23.4
# The problematic 'peft' library has been REMOVED.

# --- SpaCy Model ---
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz


python-jose[cryptography]==3.3.0
PyJWT==2.8.0

phonenumbers==8.13.0
python-dateutil==2.8.2
